Refactor project structure and resolve issues
