--- backend/data/companies-data.json
@@ -1,15 +1,0 @@
-{
-    "companies": [
-        {
-            "name": "the redstone",
-            "link": "https://www.linkedin.com/company/theredstoneai/",
-            "fbLink": "https://www.facebook.com/RealDubaians",
-            "instaLink": "https://www.instagram.com/theredstoneai/"
-        },
-        {
-            "name": "SyNet Digital",
-            "link": "https://www.linkedin.com/company/synet-digital/",
-            "fbLink": "https://www.facebook.com/SyNetDigitalMarketing"
-        }
-    ]
-}
