+++ backend/src/config/constants.ts
@@ -8,8 +8,8 @@ export const CONFIG = {
   MAX_RETRIES: 3,
   RETRY_DELAY: 2000,
   DATA_PATHS: {
-    USERS: path.join(process.cwd(), "data", "users-data.json"),
-    COMPANIES: path.join(process.cwd(), "data", "companies-data.json"),
+    USERS: path.join(process.cwd(), "src","data", "users-data.json"),
+    COMPANIES: path.join(process.cwd(), "src","data", "companies-data.json"),
     LOGS_DIR: path.join(__dirname, "..", "botLogs"),
   },
   ERROR_INDICATORS: ["Error", "timeout of", "ERROR", "crashed after", "Session ended", "Breaking forever", "Stopped", "Manually stopped", "Captcha/Code", "IP Config", "paused"],
