+++ backend/src/data/users-data.json
@@ -1,0 +1,22 @@
+{
+  "users": [
+    {
+      "username": "afaq.ahmed10009@gmail.com",
+      "password": "mySelf123$$",
+      "platforms": [ "linkedin" ],
+      "ip_address": "",
+      "ip_port": "",
+      "ip_username": "",
+      "ip_password": ""
+    },
+    {
+      "username": "attockofficeworkonly@gmail.com",
+      "password": "Makhdoom123$$",
+      "platforms": [ "linkedin", "instagram", "facebook" ],
+      "ip_address": "",
+      "ip_port": "",
+      "ip_username": "",
+      "ip_password": ""
+    }
+  ]
+}
