+++ frontend/.env.example
@@ -1,2 +1,2 @@
 # do not add ' / ' at the end of backend url
-REACT_APP_API_URL=http://localhost:8080
+REACT_APP_API_URL=http://localhost:8080/api
